+++ src/eve_overview_pro/__init__.py
@@ -1,2 +1,2 @@
-"""EVE Veles Eyes v2.3 - ActionRegistry Edition"""
+"""Argus Overview v2.3 - ActionRegistry Edition"""
 __version__ = "2.3.0"